+++ tests/test_composites.py
@@ -11,6 +11,8 @@
 import py_trees
 import py_trees.console as console
 
+from nose.tools import assert_raises
+
 ##############################################################################
 # Logging Level
 ##############################################################################
@@ -19,8 +21,32 @@ py_trees.logging.level = py_trees.logging.Level.DEBUG
 logger = py_trees.logging.Logger("Nosetest")
 
 ##############################################################################
+# Classes
+##############################################################################
+
+class InvalidSetup(py_trees.behaviour.Behaviour):
+    def setup(self, timeout):
+        # A common mistake is to forget to return a boolean value
+        # Composite behaviours will at least check to make sure that
+        # their children do so and raise TypeError's if they fail
+        # to do so.
+        pass
+
+##############################################################################
 # Tests
 ##############################################################################
+
+def test_invalid_setup():
+    console.banner("Invalid Setup")
+    parent = py_trees.composites.Sequence(name="Parent")
+    child = InvalidSetup(name="Invalid Setup")
+    parent.add_child(child)
+    print("\n--------- Assertions ---------\n")
+    print("TypeError is raised")
+    with assert_raises(TypeError) as context:
+        parent.setup(timeout=15)
+    print("TypeError has message with substring 'NoneType'")
+    assert("NoneType" in str(context.exception))
 
 
 def test_replacing_children():